Adapt approach to honggfuzz, avoid allocation crashes
This commit adjust the fuzzer from AFL to honggfuzz. Honggfuzz has the
advantage of being multi-threaded by default but, honestly, the
initial big draw was easy hook-in to lldb. The AFL fuzzer from the
previous commit 'detected' HashMap panicing when we requested too much
capacity but I found that an easy call to "cargo hfuzz run-debug
hash_map" made the failure much more clear.

The main change here is around the Reserve operation, limited now to a
smallish reservation bump. This avoids the case where HashMap will
panic on call to its `reserve`, being unable to signal that its
allocation failed. The model is also careful about requested too much
capacity, as noted. If HashMap published its overhead per pair we
could calculate when the allocations would fail but it, reasonably,
does not publish such information.

+++ src/stdlib/collections/hash_map.rs
@@ -159,30 +159,36 @@ where
 /// step.
 #[derive(Clone, Debug)]
 pub enum Op<K, V> {
-    /// TODO
+    /// This opertion triggers `std::collections::HashMap::shrink_to_fit`
     ShrinkToFit,
-    /// TODO
+    /// This operation triggers `std::collections::HashMap::clear`
     Clear,
-    /// TODO
+    /// This operation triggers `std::collections::HashMap::reserve`
     Reserve {
-        /// TODO
-        n: usize,
+        /// Reserve `n` capacity elements
+        ///
+        /// Why is this not usize? It's possible we'll try to reserve
+        /// too much underlying capacity, forcing the `reserve` call to panic
+        /// when the underlying allocation fails. The HashMap does not document
+        /// its overhead per pair and `try_reserve` is still behind a feature
+        /// flag. So, we only reserve smallish capacity and hope for the best.
+        n: u16,
     },
-    /// TODO
+    /// This operation triggers `std::collections::HashMap::insert`
     Insert {
-        /// TODO
+        /// The key to be inserted
         k: K,
-        /// TODO
+        /// The value to be inserted
         v: V,
     },
-    /// TODO
+    /// This operation triggers `std::collections::HashMap::remove`
     Remove {
-        /// TODO
+        /// The key to be removed
         k: K,
     },
-    /// TODO
+    /// This operation triggers `std::collections::HashMap::get`
     Get {
-        /// TODO
+        /// The key to be removed
         k: K,
     },
 }
@@ -221,7 +227,7 @@ where
             3 => Op::ShrinkToFit,
             4 => Op::Clear,
             5 => {
-                let n: usize = Arbitrary::arbitrary(u)?;
+                let n: u16 = Arbitrary::arbitrary(u)?;
                 Op::Reserve { n }
             }
             _ => unreachable!(),
